Refactor: chatting sidebar view, container 분리 #108
